requirements.txt
flask
pyusb
requests